refactor(web): share one meeting-row renderer across cart and search
Extract the cart's status-aware meeting row into MeetingRowCard and render search results through it via sectionSignature, replacing the duplicated markup and the second dedupe path (getUniqueMeetings, now removed). Meeting-row keys drop their redundant status prefixes.
